user-specific location and some styling using semantic ui

+++ src/components/member/MemberDetail.js
@@ -14,32 +14,39 @@ export const MemberDetail = () => {
       setMember(response);
     });
   }, []);
+
+  const showButton = () => {
+    if (member.userId === parseInt(localStorage.getItem("active_user")))
+      return (
+        <>
+          <button
+            onClick={() => {
+              deleteMember(member.id).then(() => {
+                history.push("/members");
+              });
+            }}
+          >
+            Terminate Membership
+          </button>
+          <button
+            onClick={() => {
+              history.push(`/members/edit/${member.id}`);
+            }}
+          >
+            Edit
+          </button>
+        </>
+      );
+  };
   return (
     <section className="member">
       <h3 className="member-name">Name: {member.username}</h3>
       <div className="member-email">Email: {member.email}</div>
-      <div className="member-password">Password: {member.password}</div>
       <div className="member-gender">Gender: {member.gender}</div>
       <div className="member-type">Admin?:{member.usertype}</div>
       <div className="member-zip">Zip-Code:{member.zipcode}</div>
       <div className="memberId">{activeUser}</div>
-
-      <button
-        onClick={() => {
-          deleteMember(member.id).then(() => {
-            history.push("/members");
-          });
-        }}
-      >
-        Terminate Membership
-      </button>
-      <button
-        onClick={() => {
-          history.push(`/members/edit/${member.id}`);
-        }}
-      >
-        Edit
-      </button>
+      <section className="buttons">{showButton()}</section>
     </section>
   );
 };
